pyproject.toml
[project]
name = "squareline-to-esphome"
version = "0.1.0"
description = "SquarelineToEsphome is a tool that converts UI projects created with SquareLine Studio into configuration files compatible with ESPHome"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow==11.2.1",
    "pyperclip==1.9.0",
    "pyyaml==6.0.2",
    "setuptools>=80.8.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=8.0.0",
]

[project.scripts]
squareline-to-esphome = "squareline_to_esphome.__main__:main"


[build-system]
requires = ["setuptools==80.8.0",]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
]
ignore = [
    "E501",  # line too long, handled by formatter
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
